refactor(tests): add testing utils file
